edu_backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.3
psycopg[binary]==3.2.6
PyJWT==2.10.1
